+++ src/auth/index.ts
@@ -1,0 +1,5 @@
+export {
+  UserClaimRequirements,
+  doesUserSatisfyClaimRequirements,
+} from './user-claims.js';
+export * from './user.js';