+++ src/nestjs/auth/index.ts
@@ -1,0 +1,4 @@
+export * from './auth-user.decorator.js';
+export * from './auth.module.js';
+export { Public } from './public.decorator.js';
+export { RequireUserClaims } from './require-user-claims.decorator.js';